zadatak dodatni

+++ Desktop/vue/cars/src/components/AppCars.vue
@@ -58,8 +58,13 @@ export default {
         async deleteCar(id) {
             
             try{ 
+                
+                if(confirm('Are you sure you want to delete this car?') == false) {
+                    return;
+                }
 
                 await CarsService.delete('/cars/' + id);
+                this.cars = this.cars.filter(car => car.id !== id);
 
             } catch(e) {
                 console.log(e);
